+++ src/typeguard/__init__.py
@@ -21,7 +21,6 @@ from ._functions import warn_on_error as warn_on_error
 from ._importhook import ImportHookManager as ImportHookManager
 from ._importhook import TypeguardFinder as TypeguardFinder
 from ._importhook import install_import_hook as install_import_hook
-from ._memo import CallMemo as CallMemo
 from ._memo import TypeCheckMemo as TypeCheckMemo
 
 # Re-export imports so they look like they live directly in this package