+++ tests/test_transformer.py
@@ -26,12 +26,12 @@ def test_arguments_only() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_argument_types
 
             def foo(x: int) -> None:
-                call_memo = CallMemo(foo, locals())
-                check_argument_types(call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                check_argument_types('foo', {'x': (x, int)}, memo)
         """
         ).strip()
     )
@@ -51,12 +51,12 @@ def test_return_only() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_return_type
 
             def foo(x) -> int:
-                call_memo = CallMemo(foo, locals())
-                return check_return_type(6, call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                return check_return_type('foo', 6, int, memo)
         """
         ).strip()
     )
@@ -68,6 +68,7 @@ class TestGenerator:
             dedent(
                 """
                 from collections.abc import Generator
+                from typing import Any
 
                 def foo(x) -> Generator[int, Any, str]:
                     yield 2
@@ -81,16 +82,16 @@ class TestGenerator:
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
-                from typeguard._functions import check_return_type, check_send_type, \
-check_yield_type
+                from typeguard import TypeCheckMemo
+                from typeguard._functions import check_return_type, check_yield_type
                 from collections.abc import Generator
+                from typing import Any
 
                 def foo(x) -> Generator[int, Any, str]:
-                    call_memo = CallMemo(foo, locals())
-                    check_send_type((yield check_yield_type(2, call_memo)), call_memo)
-                    check_send_type((yield check_yield_type(6, call_memo)), call_memo)
-                    return check_return_type('test', call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    yield check_yield_type('foo', 2, int, memo)
+                    yield check_yield_type('foo', 6, int, memo)
+                    return check_return_type('foo', 'test', str, memo)
             """
             ).strip()
         )
@@ -112,14 +113,16 @@ check_yield_type
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_send_type, check_yield_type
                 from collections.abc import Generator
 
                 def foo(x) -> Generator[int, None, None]:
-                    call_memo = CallMemo(foo, locals())
-                    check_send_type((yield check_yield_type(2, call_memo)), call_memo)
-                    check_send_type((yield check_yield_type(6, call_memo)), call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_send_type('foo', (yield check_yield_type('foo', 2, int, \
+memo)), None, memo)
+                    check_send_type('foo', (yield check_yield_type('foo', 6, int, \
+memo)), None, memo)
             """
             ).strip()
         )
@@ -142,15 +145,15 @@ check_yield_type
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_yield_type
                 from typing import Any
                 from collections.abc import Generator
 
                 def foo(x) -> Generator[int, Any, Any]:
-                    call_memo = CallMemo(foo, locals())
-                    yield check_yield_type(2, call_memo)
-                    yield check_yield_type(6, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    yield check_yield_type('foo', 2, int, memo)
+                    yield check_yield_type('foo', 6, int, memo)
             """
             ).strip()
         )
@@ -174,14 +177,16 @@ class TestAsyncGenerator:
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_send_type, check_yield_type
                 from collections.abc import AsyncGenerator
 
                 async def foo(x) -> AsyncGenerator[int, None]:
-                    call_memo = CallMemo(foo, locals())
-                    check_send_type((yield check_yield_type(2, call_memo)), call_memo)
-                    check_send_type((yield check_yield_type(6, call_memo)), call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_send_type('foo', (yield check_yield_type('foo', 2, int, \
+memo)), None, memo)
+                    check_send_type('foo', (yield check_yield_type('foo', 6, int, \
+memo)), None, memo)
             """
             ).strip()
         )
@@ -204,15 +209,15 @@ class TestAsyncGenerator:
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_send_type
                 from typing import Any
                 from collections.abc import AsyncGenerator
 
                 async def foo() -> AsyncGenerator[Any, None]:
-                    call_memo = CallMemo(foo, locals())
-                    check_send_type((yield 2), call_memo)
-                    check_send_type((yield 6), call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_send_type('foo', (yield 2), None, memo)
+                    check_send_type('foo', (yield 6), None, memo)
                 """
             ).strip()
         )
@@ -235,15 +240,15 @@ class TestAsyncGenerator:
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_yield_type
                 from typing import Any
                 from collections.abc import AsyncGenerator
 
                 async def foo() -> AsyncGenerator[int, Any]:
-                    call_memo = CallMemo(foo, locals())
-                    yield check_yield_type(2, call_memo)
-                    yield check_yield_type(6, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    yield check_yield_type('foo', 2, int, memo)
+                    yield check_yield_type('foo', 6, int, memo)
                 """
             ).strip()
         )
@@ -347,10 +352,10 @@ def test_avoid_global_names() -> None:
     node = parse(
         dedent(
             """
-            call_memo = CallMemo = check_argument_types = check_return_type = None
+            memo = TypeCheckMemo = check_argument_types = check_return_type = None
 
             def func1(x: int) -> int:
-                dummy = (call_memo,)
+                dummy = (memo,)
                 return x
 
             def func2(x: int) -> int:
@@ -363,21 +368,21 @@ def test_avoid_global_names() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo as CallMemo_
+            from typeguard import TypeCheckMemo as TypeCheckMemo_
             from typeguard._functions import \
 check_argument_types as check_argument_types_, check_return_type as check_return_type_
-            call_memo = CallMemo = check_argument_types = check_return_type = None
+            memo = TypeCheckMemo = check_argument_types = check_return_type = None
 
             def func1(x: int) -> int:
-                call_memo_ = CallMemo_(func1, locals())
-                check_argument_types_(call_memo_)
-                dummy = (call_memo,)
-                return check_return_type_(x, call_memo_)
+                memo_ = TypeCheckMemo_(globals(), locals())
+                check_argument_types_('func1', {'x': (x, int)}, memo_)
+                dummy = (memo,)
+                return check_return_type_('func1', x, int, memo_)
 
             def func2(x: int) -> int:
-                call_memo_ = CallMemo_(func2, locals())
-                check_argument_types_(call_memo_)
-                return check_return_type_(x, call_memo_)
+                memo_ = TypeCheckMemo_(globals(), locals())
+                check_argument_types_('func2', {'x': (x, int)}, memo_)
+                return check_return_type_('func2', x, int, memo_)
         """
         ).strip()
     )
@@ -388,7 +393,7 @@ def test_avoid_local_names() -> None:
         dedent(
             """
             def foo(x: int) -> int:
-                call_memo = CallMemo = check_argument_types = check_return_type = None
+                memo = TypeCheckMemo = check_argument_types = check_return_type = None
                 return x
             """
         )
@@ -399,13 +404,13 @@ def test_avoid_local_names() -> None:
         == dedent(
             """
             def foo(x: int) -> int:
-                from typeguard import CallMemo as CallMemo_
+                from typeguard import TypeCheckMemo as TypeCheckMemo_
                 from typeguard._functions import \
 check_argument_types as check_argument_types_, check_return_type as check_return_type_
-                call_memo_ = CallMemo_(foo, locals())
-                check_argument_types_(call_memo_)
-                call_memo = CallMemo = check_argument_types = check_return_type = None
-                return check_return_type_(x, call_memo_)
+                memo_ = TypeCheckMemo_(globals(), locals())
+                check_argument_types_('foo', {'x': (x, int)}, memo_)
+                memo = TypeCheckMemo = check_argument_types = check_return_type = None
+                return check_return_type_('foo', x, int, memo_)
             """
         ).strip()
     )
@@ -416,7 +421,7 @@ def test_avoid_nonlocal_names() -> None:
         dedent(
             """
             def outer():
-                call_memo = CallMemo = check_argument_types = check_return_type = None
+                memo = TypeCheckMemo = check_argument_types = check_return_type = None
 
                 def foo(x: int) -> int:
                     return x
@@ -431,15 +436,15 @@ def test_avoid_nonlocal_names() -> None:
         == dedent(
             """
             def outer():
-                call_memo = CallMemo = check_argument_types = check_return_type = None
+                memo = TypeCheckMemo = check_argument_types = check_return_type = None
 
                 def foo(x: int) -> int:
-                    from typeguard import CallMemo as CallMemo_
+                    from typeguard import TypeCheckMemo as TypeCheckMemo_
                     from typeguard._functions import \
 check_argument_types as check_argument_types_, check_return_type as check_return_type_
-                    call_memo_ = CallMemo_(foo, locals())
-                    check_argument_types_(call_memo_)
-                    return check_return_type_(x, call_memo_)
+                    memo_ = TypeCheckMemo_(globals(), locals())
+                    check_argument_types_('outer.<locals>.foo', {'x': (x, int)}, memo_)
+                    return check_return_type_('outer.<locals>.foo', x, int, memo_)
                 return foo
             """
         ).strip()
@@ -464,12 +469,12 @@ def test_method() -> None:
             class Foo:
 
                 def foo(self, x: int) -> int:
-                    from typeguard import CallMemo
+                    from typeguard import TypeCheckMemo
                     from typeguard._functions import check_argument_types, \
 check_return_type
-                    call_memo = CallMemo(Foo.foo, locals(), self.__class__)
-                    check_argument_types(call_memo)
-                    return check_return_type(x, call_memo)
+                    memo = TypeCheckMemo(globals(), locals(), self.__class__)
+                    check_argument_types('Foo.foo', {'x': (x, int)}, memo)
+                    return check_return_type('Foo.foo', x, int, memo)
             """
         ).strip()
     )
@@ -495,12 +500,12 @@ def test_classmethod() -> None:
 
                 @classmethod
                 def foo(cls, x: int) -> int:
-                    from typeguard import CallMemo
+                    from typeguard import TypeCheckMemo
                     from typeguard._functions import check_argument_types, \
 check_return_type
-                    call_memo = CallMemo(Foo.foo, locals(), cls)
-                    check_argument_types(call_memo)
-                    return check_return_type(x, call_memo)
+                    memo = TypeCheckMemo(globals(), locals(), cls)
+                    check_argument_types('Foo.foo', {'x': (x, int)}, memo)
+                    return check_return_type('Foo.foo', x, int, memo)
             """
         ).strip()
     )
@@ -526,12 +531,12 @@ def test_staticmethod() -> None:
 
                 @staticmethod
                 def foo(x: int) -> int:
-                    from typeguard import CallMemo
+                    from typeguard import TypeCheckMemo
                     from typeguard._functions import check_argument_types, \
 check_return_type
-                    call_memo = CallMemo(Foo.foo, locals())
-                    check_argument_types(call_memo)
-                    return check_return_type(x, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('Foo.foo', {'x': (x, int)}, memo)
+                    return check_return_type('Foo.foo', x, int, memo)
             """
         ).strip()
     )
@@ -560,12 +565,12 @@ def test_local_function() -> None:
             def wrapper():
 
                 def foo(x: int) -> int:
-                    from typeguard import CallMemo
+                    from typeguard import TypeCheckMemo
                     from typeguard._functions import check_argument_types, \
 check_return_type
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
-                    return check_return_type(x, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('wrapper.<locals>.foo', {'x': (x, int)}, memo)
+                    return check_return_type('wrapper.<locals>.foo', x, int, memo)
 
                 def foo2(x: int) -> int:
                     return x
@@ -605,12 +610,14 @@ def test_function_local_class_method() -> None:
                     class Bar:
 
                         def method(self, x: int) -> int:
-                            from typeguard import CallMemo
+                            from typeguard import TypeCheckMemo
                             from typeguard._functions import check_argument_types, \
 check_return_type
-                            call_memo = CallMemo(method, locals(), self.__class__)
-                            check_argument_types(call_memo)
-                            return check_return_type(x, call_memo)
+                            memo = TypeCheckMemo(globals(), locals(), self.__class__)
+                            check_argument_types('wrapper.<locals>.Foo.Bar.method', \
+{'x': (x, int)}, memo)
+                            return check_return_type(\
+'wrapper.<locals>.Foo.Bar.method', x, int, memo)
 
                         def method2(self, x: int) -> int:
                             return x
@@ -633,12 +640,12 @@ def test_keyword_only_argument() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_argument_types
 
             def foo(*, x: int) -> None:
-                call_memo = CallMemo(foo, locals())
-                check_argument_types(call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                check_argument_types('foo', {'x': (x, int)}, memo)
             """
         ).strip()
     )
@@ -658,12 +665,12 @@ def test_positional_only_argument() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_argument_types
 
             def foo(x: int, /) -> None:
-                call_memo = CallMemo(foo, locals())
-                check_argument_types(call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                check_argument_types('foo', {'x': (x, int)}, memo)
             """
         ).strip()
     )
@@ -683,12 +690,12 @@ def test_variable_positional_argument() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_argument_types
 
             def foo(*args: int) -> None:
-                call_memo = CallMemo(foo, locals())
-                check_argument_types(call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                check_argument_types('foo', {'args': (args, tuple[int, ...])}, memo)
             """
         ).strip()
     )
@@ -708,12 +715,12 @@ def test_variable_keyword_argument() -> None:
         unparse(node)
         == dedent(
             """
-            from typeguard import CallMemo
+            from typeguard import TypeCheckMemo
             from typeguard._functions import check_argument_types
 
             def foo(**kwargs: int) -> None:
-                call_memo = CallMemo(foo, locals())
-                check_argument_types(call_memo)
+                memo = TypeCheckMemo(globals(), locals())
+                check_argument_types('foo', {'kwargs': (kwargs, dict[str, int])}, memo)
             """
         ).strip()
     )
@@ -812,13 +819,13 @@ class TestAssign:
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_variable_assignment
 
                 def foo() -> None:
-                    call_memo = CallMemo(foo, locals())
+                    memo = TypeCheckMemo(globals(), locals())
                     x: int = check_variable_assignment(otherfunc(), {'x': int}, \
-call_memo)
+memo)
                 """
             ).strip()
         )
@@ -845,15 +852,16 @@ call_memo)
             unparse(node)
             == dedent(
                 f"""
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_argument_types, \
 check_variable_assignment
                 {extra_import}
                 def foo(*args: int) -> None:
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('foo', {{'args': (args, \
+{tuple_type}[int, ...])}}, memo)
                     args = check_variable_assignment((5,), \
-{{'args': {tuple_type}[int, ...]}}, call_memo)
+{{'args': {tuple_type}[int, ...]}}, memo)
                 """
             ).strip()
         )
@@ -880,15 +888,16 @@ check_variable_assignment
             unparse(node)
             == dedent(
                 f"""
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_argument_types, \
 check_variable_assignment
                 {extra_import}
                 def foo(**kwargs: int) -> None:
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('foo', {{'kwargs': (kwargs, \
+{dict_type}[str, int])}}, memo)
                     kwargs = check_variable_assignment({{'a': 5}}, \
-{{'kwargs': {dict_type}[str, int]}}, call_memo)
+{{'kwargs': {dict_type}[str, int]}}, memo)
                 """
             ).strip()
         )
@@ -910,15 +919,15 @@ check_variable_assignment
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_variable_assignment
                 from typing import Union as Union_
                 Union = None
 
                 def foo() -> None:
-                    call_memo = CallMemo(foo, locals())
+                    memo = TypeCheckMemo(globals(), locals())
                     x: Union_[int, str] = check_variable_assignment(otherfunc(), \
-{'x': Union_[int, str]}, call_memo)
+{'x': Union_[int, str]}, memo)
                 """
             ).strip()
         )
@@ -940,16 +949,16 @@ check_variable_assignment
             unparse(node)
             == dedent(
                 f"""
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_variable_assignment
                 from typing import Any
 
                 def foo() -> None:
-                    call_memo = CallMemo(foo, locals())
+                    memo = TypeCheckMemo(globals(), locals())
                     x: int
                     z: bytes
                     {target} = check_variable_assignment(otherfunc(), \
-{{'x': int, 'y': Any, 'z': bytes}}, call_memo)
+{{'x': int, 'y': Any, 'z': bytes}}, memo)
                 """
             ).strip()
         )
@@ -968,14 +977,14 @@ check_variable_assignment
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_argument_types, \
 check_variable_assignment
 
                 def foo(x: int) -> None:
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
-                    x = check_variable_assignment(6, {'x': int}, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('foo', {'x': (x, int)}, memo)
+                    x = check_variable_assignment(6, {'x': int}, memo)
                 """
             ).strip()
         )
@@ -996,14 +1005,14 @@ check_variable_assignment
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_variable_assignment
 
                 def foo() -> None:
-                    call_memo = CallMemo(foo, locals())
+                    memo = TypeCheckMemo(globals(), locals())
                     x: int
                     if (x := check_variable_assignment(otherfunc(), {'x': int}, \
-call_memo)):
+memo)):
                         pass
                 """
             ).strip()
@@ -1024,15 +1033,15 @@ call_memo)):
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_argument_types, \
 check_variable_assignment
 
                 def foo(x: int) -> None:
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('foo', {'x': (x, int)}, memo)
                     if (x := check_variable_assignment(otherfunc(), {'x': int}, \
-call_memo)):
+memo)):
                         pass
                 """
             ).strip()
@@ -1070,15 +1079,15 @@ call_memo)):
             unparse(node)
             == dedent(
                 f"""
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_variable_assignment
                 from operator import {function}
 
                 def foo() -> None:
-                    call_memo = CallMemo(foo, locals())
+                    memo = TypeCheckMemo(globals(), locals())
                     x: int
                     x = check_variable_assignment({function}(x, 6), {{'x': int}}, \
-call_memo)
+memo)
                 """
             ).strip()
         )
@@ -1119,15 +1128,15 @@ call_memo)
             unparse(node)
             == dedent(
                 """
-                from typeguard import CallMemo
+                from typeguard import TypeCheckMemo
                 from typeguard._functions import check_argument_types, \
 check_variable_assignment
                 from operator import iadd
 
                 def foo(x: int) -> None:
-                    call_memo = CallMemo(foo, locals())
-                    check_argument_types(call_memo)
-                    x = check_variable_assignment(iadd(x, 6), {'x': int}, call_memo)
+                    memo = TypeCheckMemo(globals(), locals())
+                    check_argument_types('foo', {'x': (x, int)}, memo)
+                    x = check_variable_assignment(iadd(x, 6), {'x': int}, memo)
                 """
             ).strip()
         )